requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
pyannote.audio
librosa
torch
torchaudio
numpy